Add files via upload
This is the first fully functional code for the button code. If the help button is pressed, the website displays "help requested" on the site, display the order of the request, and the number of times the student pressed the button. If the cancelled button is pressed, the text "request cancelled" is displayed  for about 3 seconds before resetting everything for the student except the number of times requested.