Add files via upload
This python code pulls data from the website where the buttons display their values and does a statistical analysis of the data that it pulls (the number of times each student had requested help. It temporarily uses the html document of the website itself instead of using the actual web server. It was made on the 2/15/219